+++ R/utils.R
@@ -422,7 +422,7 @@ add_iso <- function(x, iso, ds, level, map = c("id"), append = TRUE){
     stop("level must be 2 or 3")
   
   id_ds <- sprintf("id_%s", ds)
-  key   <- c('id',id_ds,'administrative_area_level','administrative_area_level_2','administrative_area_level_3','latitude','longitude','population')
+  key   <- c('id',id_ds,'administrative_area_level','administrative_area_level_2','administrative_area_level_3','latitude','longitude','population','foreign_key','foreign_key_type')
   map   <- c(map, key[!(key %in% map)])
   
   x <- map_data(x, map)