+++ NEWS.md
@@ -4,6 +4,10 @@
 
 - Added standardized geospatial IDs (`foreign_key` and `foreign_key_type`), e.g., FIPS codes for United States and NUTS codes for Europe.
 
+## Data Update
+
+- Added data for Belgium (level 3).
+
 # COVID19 2.3.0 (2021-04-01)
 
 ## Breaking Changes